requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.3
frozendict==2.4.6
idna==3.10
multitasking==0.0.11
numpy==1.26.4
pandas==2.3.0
pandas_ta==0.3.14b0
peewee==3.18.1
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.63
